% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-scan.R
\name{get_cran_tbl}
\alias{get_cran_tbl}
\title{Search CRAN and return a tibble}
\usage{
get_cran_tbl()
}
\value{
a tibble
}
\description{
This uses tools::CRAN_package_db() then returns a tibble with the columns:
    col_1....col_m. This function is currently internal as it is called in
    scan_cran.
}
\examples{

\dontrun{
get_cran_tbl()
}

}
