% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-scan.R
\name{scan_cran}
\alias{scan_cran}
\title{Find packages mentioning key words in description.}
\usage{
scan_cran(key_words, when = "last-month")
}
\arguments{
\item{key_words}{character vector containing text to put into grepl}

\item{when}{character vector either containing "last-day", "last-week", or
"last-month".}
}
\value{
a tibble with R package names matching some strings, and download
    counts from the last month by default.
}
\description{
Find packages mentioning key words in description.
}
\examples{

\dontrun{

scan_cran_tbl <- scan_cran("miss | imp*")

}

}
